♻️(search) switch to indexing courses instead of course runs
Indexing course runs lead to a lot of duplicate content on the
frontend. When searching for a course with its name, we were
most of the time seeing all the course runs of the same course
side by side. Indexing courses was challenging because of
filtering by course run attributes (like dates). We think we
managed to tackle it by rewriting the sorting script with
Painless.