🎨(search) improve the way indices are handled
The way indices were imported from dotted path was complicated to use
and was introducing circular dependencies as a util was imported in
the settings. After this refacto, the settings are pure strings and
accessing indices is easier because the import is abstracted. They
can by accessed via an attribute or iterated.